blog_platform_backend/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
gunicorn
dj-database-url
python-dotenv
whitenoise
djangorestframework-simplejwt
Pillow
